cmark to cmark should not escape chars
If a character is legal in markdown, it should not be escaped when
writing markdown from an internal markdown representation.  The
default behavior results in things like ampersands getting escaped
with a backslash.  The resulting markdown is now different from
the original markdown.  Reusing the CMARK_OPT_NORMALIZE (which is
unused) to cause this behavior.  It probably needs its own option.